+++ client/src/Components/Pages/Profile/DetailedSection/index.js
